+++ R/trakt.search.R
@@ -2,35 +2,46 @@
 #'
 #' \code{trakt.search} pulls show stats and returns it compactly.
 #'
-#' Search for a show with a keyword (e.g. \code{"Breaking Bad"}) and receive basic info of the
-#' first search result. It's main use is to retrieve the ids or proper show title for further use,
-#' as well as receiving a quick overview of a show.
+#' Search for a show with a keyword (e.g. \code{"Breaking Bad"}) and receive basic info of the first
+#' search result. It's main use is to retrieve the ids or proper show title for further use, as well
+#' as receiving a quick overview of a show.
 #' @param query The keyword used for the search. Will be coerced to \code{character} and
-#'  \code{URLencode}d.
-#' @param type The type of data you're looking for. Defaults to \code{show}, can also
-#'  be \code{movie}, \code{episode}, \code{person} or \code{list}.
-#' @return A \code{data.frame} containing a single search result. Hopefully the one you wanted.
-#' If no result is found, the return value is \code{list(error = "Nothing found")} and a \code{warning}
+#' \code{URLencode}d. If the query ends with a 4 digit numbers, this will be used as
+#' \code{year} parameter.
+#' @param type The type of data you're looking for. Defaults to \code{show}, can also be
+#'   \code{movie}, \code{episode}, \code{person} or \code{list}.
+#' @param year Optionally filter by year.
+#' @return A \code{data.frame} containing a single search result. Hopefully the one you wanted. If
+#'   no result is found, the return value is \code{list(error = "Nothing found")} and a
+#'   \code{warning}
 #' @export
 #' @importFrom jsonlite fromJSON
 #' @import httr
-#' @note See \href{http://docs.trakt.apiary.io/reference/search/text-query}{the trakt API docs for further info}
+#' @note See \href{http://docs.trakt.apiary.io/reference/search/text-query}{the trakt API docs for
+#'   further info}
 #' @family API, search
 #' @examples
 #' \dontrun{
 #' get_trakt_credentials() # Set required API data/headers
 #' breakingbad <- trakt.search("Breaking Bad")
 #' }
-trakt.search <- function(query, type = "show"){
+trakt.search <- function(query, type = "show", year = NULL){
   if (is.null(getOption("trakt.headers"))){
     stop("HTTP headers not set, see ?get_trakt_credentials")
+  }
+
+  # Parse query for possible year
+  if (is.null(year) & !(is.na(stringr::str_match(query, "(\\d{4})$")[1]))){
+    year  <- stringr::str_match(query, "(\\d{4})$")[1]
+    query <- stringr::str_replace(query, "(\\d{4})$", "")
+    query <- stringr::str_trim(query)
   }
 
   # Construct URL, make API call
   query    <- as.character(query) # Just to make sure…
   query    <- URLencode(query)    # URL normalization
   baseURL  <- "https://api-v2launch.trakt.tv/search"
-  url      <- paste0(baseURL, "?query=", query, "&type=", type)
+  url      <- paste0(baseURL, "?query=", query, "&type=", type, "&year=", year)
   response <- trakt.api.call(url = url, convert.datetime = F)
 
   # Check if response is empty (nothing found)