+++ NEWS.md
@@ -1,5 +1,10 @@
 ## tRakt News
 
+### v0.10.2
+
+* Add `year` parameter to `trakt.search`
+* If `query` in `trakt.search` ends with a 4 digit number, this will be used as `year` parameter and stripped from the original `query`
+
 ### v0.10.1
  
 * Add `testthat` tests